experiments/context_propagation/requirements.txt
# Experiment dependencies
openai>=1.30
pydantic>=2.0
numpy>=1.26
sentence-transformers>=3.0
pyyaml>=6.0
matplotlib>=3.8
pandas>=2.2
scipy>=1.13
